pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kosta-enhanced-word-mcp-server"
version = "2.7.13"
description = "Enhanced Word MCP server for academic research collaboration with advanced search/replace, review tools, and section management"
readme = "README.md"
# Keep license metadata without requiring a file in the repo
license = { text = "MIT" }
authors = [
    {name = "Kosta Vučković", email = "kosta@brown.edu"},
    {name = "GongRzhe", email = "gongrzhe@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "python-docx>=0.8.11",
    "mcp[cli]>=1.3.0",
    "msoffcrypto-tool>=5.4.2",
    "docx2pdf>=0.1.8",
]

[project.urls]
"Homepage" = "https://github.com/ildunari/enhanced-word"
"Bug Tracker" = "https://github.com/ildunari/enhanced-word/issues"

[tool.hatch.build.targets.wheel]
only-include = [
    "word_document_server",
]
sources = ["."]

[project.scripts]
word_mcp_server = "word_document_server.main:run_server"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end MCP stdio integration tests",
    "stress: heavy stress/performance/soak tests (nightly/manual CI)",
]
